+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Regular polygons".
 - Tests for "Horse-hyperracing hyperduals".
 - Tests for "3×N tiling".
+- Tests for "Hexagonal maze - part2".
 
 ## [1.13.0] - 2022-09-30
 ### Added